skills/mohd-mursaleen/android-automation/runtime/requirements.txt
Pillow>=12.0.0
python-dotenv>=1.0.0

# Optional development tools
# pytest>=7.0.0
# black>=23.0.0
# mypy>=1.0.0
# ruff>=0.1.0
